Remove unnecessary casts reported by Eclipse
git-svn-id: https://svn.apache.org/repos/asf/tomcat/trunk@734702 13f79535-47bb-0310-9956-ffa450edef68

+++ java/org/apache/tomcat/util/buf/HexUtils.java
@@ -131,8 +131,8 @@ public final class HexUtils {
 
 	StringBuffer sb = new StringBuffer(bytes.length * 2);
 	for (int i = 0; i < bytes.length; i++) {
-	    sb.append(convertDigit((int) (bytes[i] >> 4)));
-	    sb.append(convertDigit((int) (bytes[i] & 0x0f)));
+	    sb.append(convertDigit((bytes[i] >> 4)));
+	    sb.append(convertDigit((bytes[i] & 0x0f)));
 	}
 	return (sb.toString());
